add item-level documentation; reorganize slightly
