+++ src/nn/mod.rs
@@ -124,10 +124,10 @@ mod repeated;
 mod residual;
 mod split_into;
 mod transformer;
-mod visit_tensors;
+pub mod visit_tensors;
 
 pub use module::*;
-pub use visit_tensors::*;
+use visit_tensors::*;
 
 #[cfg(feature = "numpy")]
 pub use npz::*;
