+++ src/nn/visit_tensors.rs
@@ -1,11 +1,13 @@
+use crate::prelude::*;
 use crate::shapes::Dtype;
 
-use crate::prelude::*;
-use std::{
-    fmt::Debug,
-    string::{String, ToString},
-};
+use std::{fmt::Debug, string::String};
 
+/// A struct which can be used to iterate through all tensors in a module, or to iterate through
+/// corresponding tensors in multiple modules of the same type.
+///
+/// N specifies the number of immutable references to modules of type T this stores, while M
+/// specifies the number of mutable references.
 pub struct TensorVisitor<'a, const N: usize, const M: usize, T, F> {
     refs: [&'a T; N],
     refs_mut: [&'a mut T; M],
@@ -25,6 +27,15 @@ fn map_mut<'a, A: 'a + Debug, B: 'a + Debug, const N: usize, F: FnMut(&'a mut A)
 }
 
 impl<'a, const N: usize, const M: usize, T: Debug, F> TensorVisitor<'a, N, M, T, F> {
+    /// Creates a new TensorVisitor
+    ///
+    /// Parameters:
+    /// * refs: An array of immutable references to the module type to visit.
+    /// * refs_mut: An array of mutable references to the module type to visit.
+    /// * name: An optional prefix to each tensor's name. Passing `None` will prevent names from
+    /// being tracked, and will provide `None` to [TensorFunction::call].
+    /// * func: The [TensorFunction] that this tensor should call on each group of corresponding
+    /// tnesors in `refs` and `refs_mut`
     pub fn new(
         refs: [&'a T; N],
         refs_mut: [&'a mut T; M],
@@ -58,6 +69,11 @@ impl<'a, const N: usize, const M: usize, T: Debug, F> TensorVisitor<'a, N, M, T,
         }
     }
 
+    /// Calls the stored [TensorFunction] on all groups of corresponding tensors in the modules
+    /// stored in `refs` and `refs_mut`.
+    ///
+    /// For example, if refs consists of two references to [Linear] layers, `[&linear1, &linear2]`,
+    /// then the function will be given `[&linear1.weight, &linesr2.weight]` and `[&linear1.bias, &linesr2.bias]`.
     pub fn visit<E: Dtype, D: DeviceStorage>(self) -> Result<(), F::Err>
     where
         F: TensorFunction<N, M, E, D>,
@@ -66,6 +82,16 @@ impl<'a, const N: usize, const M: usize, T: Debug, F> TensorVisitor<'a, N, M, T,
         VisitTensorGroups::visit_groups(self)
     }
 
+    /// Creates a new TensorVisitor that holds references to a particular field of each module in
+    /// refs and refs_mut and calls [TensorVisitor::visit] on it.
+    ///
+    /// Parameters:
+    /// * imm_func: Given a reference to a module, returns an immutable reference to the
+    /// field to visit.
+    /// * mut_func: Given a reference to a module, returns a mutable reference to the
+    /// field to visit.
+    /// * name: Specifies the name of the field. This should have a dot at the end unless the
+    /// field is a tensor.
     pub fn visit_field<
         E: Dtype,
         D: DeviceStorage,
@@ -74,17 +100,19 @@ impl<'a, const N: usize, const M: usize, T: Debug, F> TensorVisitor<'a, N, M, T,
         F2: FnMut(&mut T) -> &mut T2,
     >(
         &mut self,
-        func1: F1,
-        func2: F2,
+        imm_func: F1,
+        mut_func: F2,
         name: &str,
     ) -> Result<(), F::Err>
     where
         F: TensorFunction<N, M, E, D>,
         T2: VisitTensorGroups<N, M, E, D>,
     {
-        self.map(func1, func2, name).visit()
+        self.map(imm_func, mut_func, name).visit()
     }
 
+    /// Same as [TensorVisitor::visit_field], but appends the options in `options` to each
+    /// [`&[TensorFunctionOption]`](TensorFunctionOption) passed into [TensorFunction::call].
     pub fn visit_field_with_options<
         E: Dtype,
         D: DeviceStorage,
@@ -112,17 +140,24 @@ impl<'a, const N: usize, const M: usize, T: Debug, F> TensorVisitor<'a, N, M, T,
     }
 }
 
+/// Configures the behavior of certain [TensorFunction]s.
 #[non_exhaustive]
 #[derive(Clone, Debug, PartialEq)]
 pub enum TensorFunctionOption {
+    /// Prevents tensors from being updated with [GradientUpdate].
     DisableGradientUpdate,
-    ResetParamsDistribution(f64, f64),
+    /// Tells ResetParams to initialize each tensor with a uniform random distribution,
+    /// with the first parameter as the lower bound and the second parameter as the upper bound.
+    ResetParamsUniform(f64, f64),
+    /// Tells ResetParams to initialize each tensor with ones.
     ResetParamsOnes,
 }
 
+/// A data type that can be called on a group of tensors, as provided by a [TensorVisitor]
 pub trait TensorFunction<const N: usize, const M: usize, E: Dtype, D: DeviceStorage> {
     type Err: std::fmt::Display + Debug;
 
+    /// Updates the  
     fn call<S: Shape>(
         &mut self,
         refs: [&Tensor<S, E, D>; N],
@@ -132,6 +167,10 @@ pub trait TensorFunction<const N: usize, const M: usize, E: Dtype, D: DeviceStor
     ) -> Result<(), Self::Err>;
 }
 
+/// Specifies how a [TensorVisitor] should traverse the fields of a Module.
+///
+/// Implementing this automatically implements [ResetParams], [GradientUpdate], [SaveToNpz],
+/// [LoadFromNpz], and [CountParams].
 pub trait VisitTensorGroups<const N: usize, const M: usize, E: Dtype, D: DeviceStorage>:
     Sized
 {
@@ -155,17 +194,21 @@ impl<const N: usize, const M: usize, S: Shape, E: Dtype, D: DeviceStorage>
     }
 }
 
+/// Alias trait for VisitTensorGroups<1, 0, E, D>
 pub trait VisitTensors<E: Dtype, D: DeviceStorage>: VisitTensorGroups<1, 0, E, D> + Debug {
+    /// Calls `func` on immutable references to all tensors in `self`.
     fn visit<F: TensorFunction<1, 0, E, D>>(&self, func: &mut F) -> Result<(), F::Err> {
         TensorVisitor::new([self], [], None, func).visit()
     }
 
+    /// Calls `func` on immutable references to all tensors in `self`, passing each tensor's name
+    /// prefixed by the name parameter into each call of [TensorFunction::call]
     fn visit_with_name<F: TensorFunction<1, 0, E, D>>(
         &self,
-        name: &str,
+        name: String,
         func: &mut F,
     ) -> Result<(), F::Err> {
-        TensorVisitor::new([self], [], Some(name.to_string()), func).visit()
+        TensorVisitor::new([self], [], Some(name), func).visit()
     }
 }
 
@@ -174,19 +217,23 @@ impl<E: Dtype, D: DeviceStorage, T> VisitTensors<E, D> for T where
 {
 }
 
+/// Alias trait for VisitTensorGroups<0, 1, E, D>
 pub trait VisitTensorsMut<E: Dtype, D: DeviceStorage>:
     VisitTensorGroups<0, 1, E, D> + Debug
 {
+    /// Call `func` on mutable references to all tensors in `self`.
     fn visit_mut<F: TensorFunction<0, 1, E, D>>(&mut self, func: &mut F) -> Result<(), F::Err> {
         VisitTensorGroups::visit_groups(TensorVisitor::new([], [self], None, func))
     }
 
+    /// Call `func` on mutable references to all tensors in `self`, passing each tensor's name
+    /// prefixed by the name parameter into each call of [TensorFunction.call]
     fn visit_mut_with_name<F: TensorFunction<0, 1, E, D>>(
         &mut self,
-        name: &str,
+        name: String,
         func: &mut F,
     ) -> Result<(), F::Err> {
-        TensorVisitor::new([], [self], Some(name.to_string()), func).visit()
+        TensorVisitor::new([], [self], Some(name), func).visit()
     }
 }
 
@@ -194,37 +241,6 @@ impl<E: Dtype, D: DeviceStorage, T> VisitTensorsMut<E, D> for T where
     T: VisitTensorGroups<0, 1, E, D> + Debug
 {
 }
-
-struct CountParamsVisitor(usize);
-
-impl<E: Dtype, D: DeviceStorage> TensorFunction<1, 0, E, D> for CountParamsVisitor {
-    type Err = String;
-
-    fn call<S: Shape>(
-        &mut self,
-        refs: [&Tensor<S, E, D>; 1],
-        _refs_mut: [&mut Tensor<S, E, D>; 0],
-        _name: Option<String>,
-        _options: &[TensorFunctionOption],
-    ) -> Result<(), Self::Err> {
-        self.0 += refs[0].shape().num_elements();
-        Ok(())
-    }
-}
-
-pub trait CountParams<E: Dtype, D: DeviceStorage>: VisitTensors<E, D> {
-    fn try_param_count(&self) -> Result<usize, String> {
-        let mut visitor = CountParamsVisitor(0);
-        self.visit(&mut visitor)?;
-        Ok(visitor.0)
-    }
-
-    fn param_count(&self) -> usize {
-        self.try_param_count().unwrap()
-    }
-}
-
-impl<E: Dtype, D: DeviceStorage, T: VisitTensors<E, D>> CountParams<E, D> for T {}
 
 #[cfg(test)]
 mod tests {
